Move ad hoc tests into proper tests
